+++ projects/clr/hipamd/include/hip/amd_detail/hip_api_trace.hpp
@@ -48,7 +48,7 @@
 #define HIP_API_TABLE_STEP_VERSION 0
 #define HIP_COMPILER_API_TABLE_STEP_VERSION 0
 #define HIP_TOOLS_API_TABLE_STEP_VERSION 1
-#define HIP_RUNTIME_API_TABLE_STEP_VERSION 34
+#define HIP_RUNTIME_API_TABLE_STEP_VERSION 35
 
 // HIP API interface
 // HIP compiler dispatch functions
@@ -1185,6 +1185,8 @@ typedef hipError_t (*t_hipMemGetDefaultMemPool)(hipMemPool_t* memPool, hipMemLoc
                                                 hipMemAllocationType type);
 typedef hipError_t (*t_hipModuleEnumerateFunctions)(hipFunction_t* functions,
                                                     unsigned int numFunctions, hipModule_t module);
+typedef hipError_t (*t_hipLibraryGetUnifiedFunction)(void** fptr, hipLibrary_t library,
+                                                     const char* symbol);
 // HIP Compiler dispatch table
 struct HipCompilerDispatchTable {
   // HIP_COMPILER_API_TABLE_STEP_VERSION == 0
@@ -1841,7 +1843,11 @@ struct HipDispatchTable {
   // HIP_RUNTIME_API_TABLE_STEP_VERSION == 34
   t_hipModuleEnumerateFunctions hipModuleEnumerateFunctions_fn;
 
+  // HIP_RUNTIME_API_TABLE_STEP_VERSION == 35
+  t_hipLibraryGetUnifiedFunction hipLibraryGetUnifiedFunction_fn;
+
   // DO NOT EDIT ABOVE!
+  // HIP_RUNTIME_API_TABLE_STEP_VERSION == 36
 
 
   // ******************************************************************************************* //