+++ projects/clr/hipamd/include/hip/amd_detail/hip_prof_str.h
@@ -507,7 +507,8 @@ enum hip_api_id_t {
   HIP_API_ID_hipDeviceGetLuid = 485,
   HIP_API_ID_hipInitDevice = 486,
   HIP_API_ID_hipModuleEnumerateFunctions = 487,
-  HIP_API_ID_LAST = 487,
+  HIP_API_ID_hipLibraryGetUnifiedFunction = 488,
+  HIP_API_ID_LAST = 488,
 
 
   HIP_API_ID_hipBindTexture = HIP_API_ID_NONE,
@@ -804,6 +805,7 @@ static inline const char* hip_api_name(const uint32_t id) {
     case HIP_API_ID_hipLibraryGetKernel: return "hipLibraryGetKernel";
     case HIP_API_ID_hipLibraryGetKernelCount: return "hipLibraryGetKernelCount";
     case HIP_API_ID_hipLibraryGetManaged: return "hipLibraryGetManaged";
+    case HIP_API_ID_hipLibraryGetUnifiedFunction: return "hipLibraryGetUnifiedFunction";
     case HIP_API_ID_hipLibraryLoadData: return "hipLibraryLoadData";
     case HIP_API_ID_hipLibraryLoadFromFile: return "hipLibraryLoadFromFile";
     case HIP_API_ID_hipLibraryUnload: return "hipLibraryUnload";
@@ -1285,6 +1287,7 @@ static inline uint32_t hipApiIdByName(const char* name) {
   if (strcmp("hipLibraryGetKernel", name) == 0) return HIP_API_ID_hipLibraryGetKernel;
   if (strcmp("hipLibraryGetKernelCount", name) == 0) return HIP_API_ID_hipLibraryGetKernelCount;
   if (strcmp("hipLibraryGetManaged", name) == 0) return HIP_API_ID_hipLibraryGetManaged;
+  if (strcmp("hipLibraryGetUnifiedFunction", name) == 0) return HIP_API_ID_hipLibraryGetUnifiedFunction;
   if (strcmp("hipLibraryLoadData", name) == 0) return HIP_API_ID_hipLibraryLoadData;
   if (strcmp("hipLibraryLoadFromFile", name) == 0) return HIP_API_ID_hipLibraryLoadFromFile;
   if (strcmp("hipLibraryUnload", name) == 0) return HIP_API_ID_hipLibraryUnload;
@@ -3032,6 +3035,13 @@ typedef struct hip_api_data_s {
       const char* name;
       char name__val;
     } hipLibraryGetManaged;
+    struct {
+      void** fptr;
+      void* fptr__val;
+      hipLibrary_t library;
+      const char* symbol;
+      char symbol__val;
+    } hipLibraryGetUnifiedFunction;
     struct {
       hipLibrary_t* library;
       hipLibrary_t library__val;
@@ -5951,6 +5961,12 @@ typedef struct hip_api_data_s {
   cb_data.args.hipLibraryGetManaged.library = (hipLibrary_t)library; \
   cb_data.args.hipLibraryGetManaged.name = (name) ? strdup(name) : NULL; \
 };
+// hipLibraryGetUnifiedFunction[('void**', 'fptr'), ('hipLibrary_t', 'library'), ('const char*', 'symbol')]
+#define INIT_hipLibraryGetUnifiedFunction_CB_ARGS_DATA(cb_data) { \
+  cb_data.args.hipLibraryGetUnifiedFunction.fptr = (void**)fptr; \
+  cb_data.args.hipLibraryGetUnifiedFunction.library = (hipLibrary_t)library; \
+  cb_data.args.hipLibraryGetUnifiedFunction.symbol = (symbol) ? strdup(symbol) : NULL; \
+};
 // hipLibraryLoadData[('hipLibrary_t*', 'library'), ('const void*', 'code'), ('hipJitOption*', 'jitOptions'), ('void**', 'jitOptionsValues'), ('unsigned int', 'numJitOptions'), ('hipLibraryOption*', 'libraryOptions'), ('void**', 'libraryOptionValues'), ('unsigned int', 'numLibraryOptions')]
 #define INIT_hipLibraryLoadData_CB_ARGS_DATA(cb_data) { \
   cb_data.args.hipLibraryLoadData.library = (hipLibrary_t*)library; \
@@ -8475,6 +8491,11 @@ static inline void hipApiArgsInit(hip_api_id_t id, hip_api_data_t* data) {
       if (data->args.hipLibraryGetManaged.dptr) data->args.hipLibraryGetManaged.dptr__val = *(data->args.hipLibraryGetManaged.dptr);
       if (data->args.hipLibraryGetManaged.bytes) data->args.hipLibraryGetManaged.bytes__val = *(data->args.hipLibraryGetManaged.bytes);
       if (data->args.hipLibraryGetManaged.name) data->args.hipLibraryGetManaged.name__val = *(data->args.hipLibraryGetManaged.name);
+      break;
+// hipLibraryGetUnifiedFunction[('void**', 'fptr'), ('hipLibrary_t', 'library'), ('const char*', 'symbol')]
+    case HIP_API_ID_hipLibraryGetUnifiedFunction:
+      if (data->args.hipLibraryGetUnifiedFunction.fptr) data->args.hipLibraryGetUnifiedFunction.fptr__val = *(data->args.hipLibraryGetUnifiedFunction.fptr);
+      if (data->args.hipLibraryGetUnifiedFunction.symbol) data->args.hipLibraryGetUnifiedFunction.symbol__val = *(data->args.hipLibraryGetUnifiedFunction.symbol);
       break;
 // hipLibraryLoadData[('hipLibrary_t*', 'library'), ('const void*', 'code'), ('hipJitOption*', 'jitOptions'), ('void**', 'jitOptionsValues'), ('unsigned int', 'numJitOptions'), ('hipLibraryOption*', 'libraryOptions'), ('void**', 'libraryOptionValues'), ('unsigned int', 'numLibraryOptions')]
     case HIP_API_ID_hipLibraryLoadData:
@@ -11387,6 +11408,15 @@ static inline const char* hipApiString(hip_api_id_t id, const hip_api_data_t* da
       oss << ", library="; roctracer::hip_support::detail::operator<<(oss, data->args.hipLibraryGetManaged.library);
       if (data->args.hipLibraryGetManaged.name == NULL) oss << ", name=NULL";
       else { oss << ", name="; roctracer::hip_support::detail::operator<<(oss, data->args.hipLibraryGetManaged.name__val); }
+      oss << ")";
+    break;
+    case HIP_API_ID_hipLibraryGetUnifiedFunction:
+      oss << "hipLibraryGetUnifiedFunction(";
+      if (data->args.hipLibraryGetUnifiedFunction.fptr == NULL) oss << "fptr=NULL";
+      else { oss << "fptr="; roctracer::hip_support::detail::operator<<(oss, data->args.hipLibraryGetUnifiedFunction.fptr__val); }
+      oss << ", library="; roctracer::hip_support::detail::operator<<(oss, data->args.hipLibraryGetUnifiedFunction.library);
+      if (data->args.hipLibraryGetUnifiedFunction.symbol == NULL) oss << ", symbol=NULL";
+      else { oss << ", symbol="; roctracer::hip_support::detail::operator<<(oss, data->args.hipLibraryGetUnifiedFunction.symbol__val); }
       oss << ")";
     break;
     case HIP_API_ID_hipLibraryLoadData: