+++ projects/rocprofiler-sdk/source/include/rocprofiler-sdk/hip/api_args.h
@@ -3648,6 +3648,14 @@ typedef union rocprofiler_hip_api_args_t
         hipModule_t    mod;
     } hipModuleEnumerateFunctions;
 #endif
+#if HIP_RUNTIME_API_TABLE_STEP_VERSION >= 35
+    struct
+    {
+        void**       fptr;
+        hipLibrary_t library;
+        const char*  symbol;
+    } hipLibraryGetUnifiedFunction;
+#endif
 } rocprofiler_hip_api_args_t;
 
 ROCPROFILER_EXTERN_C_FINI